+++ src/common/dao/replication_job.go
@@ -104,29 +104,18 @@ func FilterRepTargets(name string) ([]*models.RepTarget, error) {
 
 // AddRepPolicy ...
 func AddRepPolicy(policy models.RepPolicy) (int64, error) {
-	o := GetOrmer()
-	sql := `insert into replication_policy (name, project_id, target_id, enabled, description, cron_str, start_time, creation_time, update_time ) values (?, ?, ?, ?, ?, ?, ?, ?, ?)`
-	p, err := o.Raw(sql).Prepare()
-	if err != nil {
+	if err := policy.MarshalFilter(); err != nil {
 		return 0, err
 	}
-
-	params := []interface{}{}
-	params = append(params, policy.Name, policy.ProjectID, policy.TargetID, policy.Enabled, policy.Description, policy.CronStr)
 	now := time.Now()
+	policy.CreationTime = now
+	policy.UpdateTime = now
 	if policy.Enabled == 1 {
-		params = append(params, now)
-	} else {
-		params = append(params, nil)
+		policy.StartTime = now
 	}
-	params = append(params, now, now)
+	policy.Deleted = 0
 
-	r, err := p.Exec(params...)
-	if err != nil {
-		return 0, err
-	}
-	id, err := r.LastInsertId()
-	return id, err
+	return GetOrmer().Insert(&policy)
 }
 
 // GetRepPolicy ...
@@ -143,6 +132,10 @@ func GetRepPolicy(id int64) (*models.RepPolicy, error) {
 		return nil, err
 	}
 
+	if err := policy.UnmarshalFilter(); err != nil {
+		return nil, err
+	}
+
 	return &policy, nil
 }
 
@@ -154,7 +147,8 @@ func FilterRepPolicies(name string, projectID int64) ([]*models.RepPolicy, error
 
 	sql := `select rp.id, rp.project_id, rp.target_id, 
 				rt.name as target_name, rp.name, rp.enabled, rp.description,
-				rp.cron_str, rp.start_time, rp.creation_time, rp.update_time, 
+				rp.cron_str, rp.filters, rp.replicate_deletion,rp.start_time, 
+				rp.creation_time, rp.update_time, 
 				count(rj.status) as error_job_count 
 			from replication_policy rp 
 			left join replication_target rt on rp.target_id=rt.id 
@@ -180,6 +174,13 @@ func FilterRepPolicies(name string, projectID int64) ([]*models.RepPolicy, error
 	if _, err := o.Raw(sql, args).QueryRows(&policies); err != nil {
 		return nil, err
 	}
+
+	for _, policy := range policies {
+		if err := policy.UnmarshalFilter(); err != nil {
+			return nil, err
+		}
+	}
+
 	return policies, nil
 }
 
@@ -197,6 +198,10 @@ func GetRepPolicyByName(name string) (*models.RepPolicy, error) {
 		return nil, err
 	}
 
+	if err := policy.UnmarshalFilter(); err != nil {
+		return nil, err
+	}
+
 	return &policy, nil
 }
 
@@ -209,6 +214,12 @@ func GetRepPolicyByProject(projectID int64) ([]*models.RepPolicy, error) {
 
 	if _, err := o.Raw(sql, projectID).QueryRows(&policies); err != nil {
 		return nil, err
+	}
+
+	for _, policy := range policies {
+		if err := policy.UnmarshalFilter(); err != nil {
+			return nil, err
+		}
 	}
 
 	return policies, nil
@@ -225,6 +236,12 @@ func GetRepPolicyByTarget(targetID int64) ([]*models.RepPolicy, error) {
 		return nil, err
 	}
 
+	for _, policy := range policies {
+		if err := policy.UnmarshalFilter(); err != nil {
+			return nil, err
+		}
+	}
+
 	return policies, nil
 }
 
@@ -239,14 +256,24 @@ func GetRepPolicyByProjectAndTarget(projectID, targetID int64) ([]*models.RepPol
 		return nil, err
 	}
 
+	for _, policy := range policies {
+		if err := policy.UnmarshalFilter(); err != nil {
+			return nil, err
+		}
+	}
+
 	return policies, nil
 }
 
 // UpdateRepPolicy ...
 func UpdateRepPolicy(policy *models.RepPolicy) error {
+	if err := policy.MarshalFilter(); err != nil {
+		return err
+	}
 	o := GetOrmer()
 	policy.UpdateTime = time.Now()
-	_, err := o.Update(policy, "TargetID", "Name", "Enabled", "Description", "CronStr", "UpdateTime")
+	_, err := o.Update(policy, "TargetID", "Name", "Enabled", "Description",
+		"Trigger", "FiltersInDB", "ReplicateDeletion", "UpdateTime")
 	return err
 }